refactored controllers
