+++ src/main/java/edu/unifi/controller/RoomToolController.java
@@ -5,6 +5,7 @@ import edu.unifi.model.entities.Table;
 import edu.unifi.model.orm.dao.CheckDAO;
 import edu.unifi.model.orm.dao.RoomDAO;
 import edu.unifi.model.orm.dao.TableDAO;
+import edu.unifi.view.RoomCreationTool;
 import edu.unifi.view.RoomUpdateTool;
 
 import java.awt.event.ActionEvent;
@@ -12,13 +13,43 @@ import java.awt.event.ActionListener;
 import java.util.List;
 import java.util.Observable;
 
-public class RoomEditDeletionToolController {
+public class RoomToolController {
     private static List<Room> rooms;
 
     public List<Room> getFilteredRooms(String filter) {
         rooms = RoomDAO.getInstance().getAll();
         return rooms.stream().filter(room -> room.getName().toLowerCase().contains(filter.toLowerCase())).toList();
     }
+
+    public static class RoomCreationToolController extends Observable implements ActionListener {
+
+        private final RoomCreationTool roomCreationTool;
+
+        public RoomCreationToolController(RoomCreationTool roomCreationTool) {
+            this.roomCreationTool = roomCreationTool;
+        }
+
+        @Override
+        public void actionPerformed(ActionEvent e) {
+
+            Room room = new Room("");
+            String roomName = roomCreationTool.getNameTextField().getText();
+            if (roomName.isEmpty() || RoomDAO.getInstance().getByName(roomName) != null) {
+                setChanged();
+                notifyObservers(Notifier.Message.build(MessageType.ERROR, "Invalid room name"));
+                return;
+            }
+
+            room.setName(roomName);
+            RoomDAO.getInstance().insert(room);
+
+            setChanged();
+            notifyObservers(Notifier.Message.build(MessageType.ADD_ROOM, roomName + " successfully added!"));
+            roomCreationTool.dispose();
+        }
+
+    }
+
 
     public static class RoomEditController extends Observable implements ActionListener {
         private Room room;
