+++ src/main/java/edu/unifi/view/TableUpdateTool.java
@@ -2,7 +2,8 @@ package edu.unifi.view;
 
 import edu.unifi.Notifier;
 import edu.unifi.controller.CheckController;
-import edu.unifi.controller.TableUpdateController;
+import edu.unifi.controller.TableToolController;
+import edu.unifi.controller.TableToolController.TableUpdateToolController;
 import edu.unifi.model.entities.Order;
 import edu.unifi.model.entities.Table;
 import org.kordamp.ikonli.materialdesign2.MaterialDesignP;
@@ -33,9 +34,6 @@ public class TableUpdateTool extends TableCreationTool {
     private JPanel listPanel;
     private JPanel bottomPanel;
     private JLabel receiptTotalLabel;
-
-    private TableUpdateController tableController;
-
     private java.util.List<Order> orders = new ArrayList<>();
     private java.util.List<OrderListItem> orderItems = new ArrayList<>();
 
@@ -58,7 +56,7 @@ public class TableUpdateTool extends TableCreationTool {
         createButton.setText("Update");
         createButton.setIcon(FontIcon.of(MaterialDesignU.UPDATE, 20));
 
-        tableController = new TableUpdateController(table, this);
+        TableToolController.TableUpdateToolController tableController = new TableToolController.TableUpdateToolController(table, this);
         tableController.addObserver(Notifier.getInstance());
         createButton.addActionListener(tableController);
 
@@ -279,7 +277,7 @@ public class TableUpdateTool extends TableCreationTool {
     public void buildOrdersList(Table table) {
         double total = 0.00;
 
-        orders = tableController.getTableOrders(table);
+        orders = new TableToolController.TableUpdateToolController(table, this).getTableOrders(table);
         listPanel = new JPanel();
         listPanel.setLayout(new GridLayout(orders.size(), 1));
         listScroller.setViewportView(listPanel);
